+++ index.html
@@ -35,7 +35,7 @@
     <hr />
     <section id="movies_container"></section>
   </div>
-  <script src="src/movie_collection.js"></script>
+  <script type=module" src="src/movie_collection.js"></script>
 
 </body>
