+++ src/movie_collection.js
@@ -1,13 +1,6 @@
-// API키
-const API_KEY = "485a7d173048113813de904df9f34a7f";
-const options = {
-  method: "GET",
-  headers: {
-    accept: "application/json",
-    Authorization:
-      "Bearer eyJhbGciOiJIUzI1NiJ9.eyJhdWQiOiI0ODVhN2QxNzMwNDgxMTM4MTNkZTkwNGRmOWYzNGE3ZiIsIm5iZiI6MTcyMTg2NzkxNS4wNjczNywic3ViIjoiNjVmZDRjYWQ3NzA3MDAwMTdjMGE4MTZmIiwic2NvcGVzIjpbImFwaV9yZWFkIl0sInZlcnNpb24iOjF9.cbkZQGyQg8TkXQEW93ZyqtaR4dVrGhw2j-1FJPufuus",
-  }
-}
+// API키와 옵션 임포트
+import { defanikey } from "./apikey.js";
+const { API_KEY, options } = defanikey;
 
 // 영화 데이터를 저장할 전역 배열
 let $movies = [];
@@ -44,7 +37,7 @@ function displayMovies() {
 
   // 각 영화 객체에 대해 반복함(forEach 사용)
   $movies.forEach((movie) => {
-    //새로운 div요소를 생성하여 카드를 만듬
+    // 새로운 div 요소를 생성하여 카드를 만듬
     const CARD = document.createElement("div");
     // CARD div 요소에 클래스네임 추가
     CARD.classList.add("card");
@@ -71,7 +64,7 @@ function displayMovies() {
     RATING.innerText = `평점: ${movie.vote_average}`;
     CARD.appendChild(RATING);
 
-    //카드 클릭 이벤트 추가
+    // 카드 클릭 이벤트 추가
     CARD.addEventListener("click", () => {
       alert(`영화 ID: ${movie.id}`);
     });
